+++ account_move_tax_abb/README.rst
@@ -1,6 +1,6 @@
-=============================
-Accounting - Full Tax Invoice
-=============================
+====================
+Accounting - Tax ABB
+====================
 
 .. 
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
@@ -17,7 +17,7 @@ Accounting - Full Tax Invoice
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-ecosoft--odoo%2Fecosoft--addons-lightgray.png?logo=github
-    :target: https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_full_tax_invoice
+    :target: https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_tax_abb
     :alt: ecosoft-odoo/ecosoft-addons
 
 |badge1| |badge2| |badge3|
@@ -40,7 +40,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/ecosoft-odoo/ecosoft-addons/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-`feedback <https://github.com/ecosoft-odoo/ecosoft-addons/issues/new?body=module:%20account_move_full_tax_invoice%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/ecosoft-odoo/ecosoft-addons/issues/new?body=module:%20account_move_tax_abb%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -68,6 +68,6 @@ Current maintainer:
 
 |maintainer-Saran440| 
 
-This module is part of the `ecosoft-odoo/ecosoft-addons <https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_full_tax_invoice>`_ project on GitHub.
+This module is part of the `ecosoft-odoo/ecosoft-addons <https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_tax_abb>`_ project on GitHub.
 
 You are welcome to contribute.
