+++ account_move_tax_abb/__init__.py
@@ -1,4 +1,5 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import models
-from . import wizards
+
+# from . import wizards
