+++ account_move_tax_abb/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils: https://docutils.sourceforge.io/" />
-<title>Accounting - Full Tax Invoice</title>
+<title>Accounting - Tax ABB</title>
 <style type="text/css">
 
 /*
@@ -360,8 +360,8 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="accounting-full-tax-invoice">
-<h1 class="title">Accounting - Full Tax Invoice</h1>
+<div class="document" id="accounting-tax-abb">
+<h1 class="title">Accounting - Tax ABB</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
@@ -369,7 +369,7 @@ ul.auto-toc {
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! source digest: sha256:76c8a6cd202b1340f6251cb6b431c0cb47fed763819c2e9e470eeb59ac383b69
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_full_tax_invoice"><img alt="ecosoft-odoo/ecosoft-addons" src="https://img.shields.io/badge/github-ecosoft--odoo%2Fecosoft--addons-lightgray.png?logo=github" /></a></p>
+<p><a class="reference external image-reference" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external image-reference" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external image-reference" href="https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_tax_abb"><img alt="ecosoft-odoo/ecosoft-addons" src="https://img.shields.io/badge/github-ecosoft--odoo%2Fecosoft--addons-lightgray.png?logo=github" /></a></p>
 <p>TODO</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
@@ -393,7 +393,7 @@ ul.auto-toc {
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/ecosoft-odoo/ecosoft-addons/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us to smash it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/ecosoft-odoo/ecosoft-addons/issues/new?body=module:%20account_move_full_tax_invoice%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/ecosoft-odoo/ecosoft-addons/issues/new?body=module:%20account_move_tax_abb%0Aversion:%2016.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -414,7 +414,7 @@ If you spotted it first, help us to smash it by providing a detailed and welcome
 <h2><a class="toc-backref" href="#toc-entry-6">Maintainers</a></h2>
 <p>Current maintainer:</p>
 <p><a class="reference external image-reference" href="https://github.com/Saran440"><img alt="Saran440" src="https://github.com/Saran440.png?size=40px" /></a></p>
-<p>This module is part of the <a class="reference external" href="https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_full_tax_invoice">ecosoft-odoo/ecosoft-addons</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/ecosoft-odoo/ecosoft-addons/tree/16.0/account_move_tax_abb">ecosoft-odoo/ecosoft-addons</a> project on GitHub.</p>
 <p>You are welcome to contribute.</p>
 </div>
 </div>
