--- setup/account_move_full_tax_invoice/odoo/addons/account_move_full_tax_invoice
@@ -1,1 +1,0 @@
-../../../../account_move_full_tax_invoice
