+++ setup/account_move_tax_abb/odoo/addons/account_move_tax_abb
@@ -1,0 +1,1 @@
+../../../../account_move_tax_abb
